+++ src/test/extension.test.ts
@@ -1,15 +1,16 @@
 import * as assert from 'assert';
-
-// You can import and use all API from the 'vscode' module
-// as well as import your extension to test it
 import * as vscode from 'vscode';
-// import * as myExtension from '../../extension';
 
 suite('Extension Test Suite', () => {
-	vscode.window.showInformationMessage('Start all tests.');
+  vscode.window.showInformationMessage('Start all tests.');
 
-	test('Sample test', () => {
-		assert.strictEqual(-1, [1, 2, 3].indexOf(5));
-		assert.strictEqual(-1, [1, 2, 3].indexOf(0));
-	});
+  test('Extension should be present', () => {
+    assert.ok(vscode.extensions.getExtension('your-publisher-name.tailwindcss-class-sorter'));
+  });
+
+  test('Extension commands should be registered', async () => {
+    const commands = await vscode.commands.getCommands(true);
+    assert.ok(commands.includes('tailwindcss-sorter.sortClasses'));
+    assert.ok(commands.includes('tailwindcss-sorter.sortClassesInSelection'));
+  });
 });